Move YearRange constructors to companion object
